requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
playwright==1.41.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
PyJWT==2.10.1
python-jose==3.5.0
pytest-asyncio==0.23.5
